# docs/swagger.yaml
definitions:
...
  request.AdminLogin:
    properties:
      password:
        maxLength: 20
        minLength: 8
        type: string
      user_name:
        maxLength: 20
        minLength: 8
        type: string
    type: object
  request.Block:
    properties:
      user_id:
        type: integer
    required:
    - user_id
    type: object
...
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Glamgrove is an E-commerce platform for purchasing high quality beauty
    products
  title: Go + Gin E-Commerce API Glamgrove
  version: 1.0.0
paths:
...
  /admin/coupons/list:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all coupons with pagination
      parameters:
      - description: Number of coupons per page
        in: query
        name: count
        required: true
        type: integer
      - description: Page number
        in: query
        name: page_number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of coupons
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid inputs
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: List all coupons
      tags:
      - Coupons
  /admin/dashboard/:
    get:
      consumes:
      - application/json
      description: Get details for the dashboard
      produces:
      - application/json
      responses:
        "200":
          description: succesfully recevied all records
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: error in getting dashboard details
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get dashboard details
      tags:
      - Admin Dashboard
  /admin/dashboard/filteredSalesReport:
    get:
      consumes:
      - application/json
      description: Get  sales report for a specific time period
      parameters:
      - description: Time period for the sales report
        in: path
        name: period
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sales report retrieved successfully
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Sales report could not be retrieved
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get filtered sales report
      tags:
      - Admin
...
  /admin/login/:
    post:
      consumes:
      - application/json
      description: Login to Admin account.
      parameters:
      - description: inputs
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.AdminLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Admin login
      tags:
      - Admin
...
  /admin/products/imageCrop:
    post:
      consumes:
      - application/json
      description: Crops the specified image.
      parameters:
      - description: Product image ID
        in: query
        name: product_image_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Image cropped and saved successfully
          schema:
            type: object
        "400":
          description: product_image_id is required" or "Invalid product_image_id
          schema:
            type: object
        "500":
          description: Failed to crop image" or "Failed to open image" or "Failed
            to save image
          schema:
            type: object
      summary: Crop image
      tags:
      - Images
...
  /admin/users/:
    get:
      consumes:
      - application/json
      description: Get a paginated list of users.
      parameters:
      - description: Number of users to fetch per page
        in: query
        name: count
        type: integer
      - description: Page number
        in: query
        name: page_number
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List user successful
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Missing or invalid inputs
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to get all users
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get a list of users
      tags:
      - Users
  /admin/users/block:
    post:
      consumes:
      - application/json
      description: Blocks or unblocks in admin side
      parameters:
      - description: inputs
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.Block'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully changed user block_status
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to change user block_status
          schema:
            $ref: '#/definitions/response.Response'
      summary: Block or unblock user
      tags:
      - Admin
  /admin/users/return-orders:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all pending return requests with pagination
        support.
      parameters:
      - description: Number of items per page
        in: query
        name: count
        type: integer
      - description: Page number
        in: query
        name: page_number
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Return Request List
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Missing or invalid inputs" or "Something went wrong!
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all pending return requests
      tags:
      - Return Orders
...
  /coupons/list:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all coupons with pagination
      parameters:
      - description: Number of coupons per page
        in: query
        name: count
        required: true
        type: integer
      - description: Page number
        in: query
        name: page_number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of coupons
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid inputs
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      summary: List all coupons
      tags:
      - Coupons
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"